+++ src/lib/libc_preprocessor.js
@@ -81,6 +81,10 @@ addToLibrary({
 
     // Returns a tokenized array of the given string expression, i.e. "FOO > BAR && BAZ" -> ["FOO", ">", "BAR", "&&", "BAZ"]
     // Optionally keeps whitespace as tokens to be able to reconstruct the original input string.
+    /**
+     * @param {string} exprString
+     * @param {(number|boolean)=} keepWhitespace Optional, can be omitted. Defaults to false.
+    */
     function tokenize(exprString, keepWhitespace) {
       var out = [], len = exprString.length;
       for (var i = 0; i <= len; ++i) {
@@ -109,8 +113,12 @@ addToLibrary({
     }
 
     // Expands preprocessing macros on substring str[lineStart...lineEnd]
-    function expandMacros(str, lineStart, lineEnd) {
-      if (lineEnd === undefined) lineEnd = str.length;
+    /**
+    * @param {string} str
+    * @param {number} lineStart
+    * @param {number=} lineEnd Optional, may be omitted.
+    */
+    function expandMacros(str, lineStart, lineEnd=str.length) {
       var len = str.length;
       var out = '';
       for (var i = lineStart; i < lineEnd; ++i) {
@@ -168,7 +176,7 @@ addToLibrary({
 
           if (operatorAndPriority == 13 /* parens '(' */) {
             // Find the closing parens position
-            var j = find_closing_parens_index(tokens, i);
+            j = find_closing_parens_index(tokens, i);
             if (j) {
               tokens.splice(i, j+1-i, buildExprTree(tokens.slice(i+1, j)));
               return tokens;
@@ -224,7 +232,7 @@ addToLibrary({
       if (i < 0) i = len;
 
       // Find the first non-whitespace character on the line.
-      for (var j = lineStart; j < i && isWhitespace(code, j); ++j);
+      for (var j = lineStart; j < i && isWhitespace(code, j);) ++j;
 
       // Is this a non-preprocessor directive line?
       var thisLineIsInActivePreprocessingBlock = stack[stack.length-1];
@@ -246,6 +254,14 @@ addToLibrary({
         var exprTree = buildExprTree(tokens);
         var evaluated = exprTree();
         stack.push(!!evaluated * stack[stack.length-1]);
+        break;
+      case 'elif':
+        var tokens = tokenize(expandMacros(expression, 0));
+        var exprTree = buildExprTree(tokens);
+        var evaluated = exprTree();
+        // If the previous #if / #elif block was executed, output NaN so that all further #elif and #else blocks will
+        // short to false.
+        stack[stack.length-1] = !!evaluated * (stack[stack.length-1] ? NaN : 1-stack[stack.length-1]);
         break;
       case 'ifdef': stack.push(!!defs[expression] * stack[stack.length-1]); break;
       case 'ifndef': stack.push(!defs[expression] * stack[stack.length-1]); break;