+++ ChangeLog.md
@@ -18,8 +18,14 @@ to browse the changes between the tags.
 
 See docs/process.md for more on how version tagging works.
 
-4.0.23 (in development)
+4.0.24 (in development)
 -----------------------
+- compiler-rt was updated to LLVM 21.1.8. (#26405)
+- A new `-sEXECUTABLE` setting was added which adds a #! line to the resulting
+  JavaScript and makes it executable. (#26085)
+
+4.0.23 - 01/10/26
+-----------------
 - The inconsistency of incrementing / decrementing refcounts between Wasm EH and
   Emscripten EH has been fixed. See `test_EXPORT_EXCEPTION_HANDLING_HELPERS` in
   `test_core.py` to see the usage. (#25988)
@@ -656,7 +662,7 @@ See docs/process.md for more on how version tagging works.
 - Fix the location of the dummy `.worker.js` file that is now generated as part
   of pthread builds so that is generated alongside the main JavaScript file.
   See #21701. ()
-- `-sASYNCIFY=2` is setting now deprecated, use `-sJSPI` instead.
+- `-sASYNCIFY=2` is setting now deprecated, use `-sJSPI` instead. (#21824)
 
 3.1.58 - 04/23/24
 -----------------